requirements.txt
# Web scraping
selenium==4.18.1
webdriver-manager==4.0.1

# Storage and data handling
google-cloud-storage==2.14.0
requests==2.31.0

# Optional but useful additions
pandas==2.2.1  # For data analysis
python-dotenv==1.0.1  # For environment variables 